requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==22.10.0
blinker==1.8.2
cffi==1.17.0
channels==4.1.0
click==8.1.7
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Django==5.1
Flask==3.0.3
grpcio==1.65.5
grpcio-tools==1.65.5
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.2.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
web-terminal==0.2.1
websockets==12.0
Werkzeug==3.0.3
zope.interface==7.0.1